# docs/api-contract/openapi.snapshot.json
{
  "components": {
    "schemas": {
      "ApiErrorPayload": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "default": "",
            "title": "Detail"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiErrorPayload",
        "type": "object"
      },
...
      "ApiResponse_KnowledgeAcceptanceReportData_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KnowledgeAcceptanceReportData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "trace_id": {
            "default": "",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "success"
        ],
        "title": "ApiResponse[KnowledgeAcceptanceReportData]",
        "type": "object"
      },
...
      "ApiResponse_KnowledgeImportResultData_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KnowledgeImportResultData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "trace_id": {
            "default": "",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "success"
        ],
        "title": "ApiResponse[KnowledgeImportResultData]",
        "type": "object"
      },
      "ApiResponse_KnowledgeNormalizeConfirmResultData_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KnowledgeNormalizeConfirmResultData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "trace_id": {
            "default": "",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "success"
        ],
        "title": "ApiResponse[KnowledgeNormalizeConfirmResultData]",
        "type": "object"
      },
      "ApiResponse_KnowledgeNormalizePreviewData_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KnowledgeNormalizePreviewData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "trace_id": {
            "default": "",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "success"
        ],
        "title": "ApiResponse[KnowledgeNormalizePreviewData]",
        "type": "object"
      },
...
      "ApiResponse_KnowledgeTrustedRebuildResultData_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KnowledgeTrustedRebuildResultData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "trace_id": {
            "default": "",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "success"
        ],
        "title": "ApiResponse[KnowledgeTrustedRebuildResultData]",
        "type": "object"
      },
...
      "ApiResponse_WebKnowledgeBuildResultData_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WebKnowledgeBuildResultData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "trace_id": {
            "default": "",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "success"
        ],
        "title": "ApiResponse[WebKnowledgeBuildResultData]",
        "type": "object"
      },
...
      "ApiResponse_list_KnowledgeSearchResultData__": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/KnowledgeSearchResultData"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "trace_id": {
            "default": "",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "success"
        ],
        "title": "ApiResponse[list[KnowledgeSearchResultData]]",
        "type": "object"
      },
      "ApiResponse_list_KnowledgeTaskData__": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/KnowledgeTaskData"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "trace_id": {
            "default": "",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "success"
        ],
        "title": "ApiResponse[list[KnowledgeTaskData]]",
        "type": "object"
      },
...
      "Body_upload_knowledge_api_v1_knowledge_upload_post": {
        "properties": {
          "files": {
            "items": {
              "contentMediaType": "application/octet-stream",
              "type": "string"
            },
            "title": "Files",
            "type": "array"
          }
        },
        "required": [
          "files"
        ],
        "title": "Body_upload_knowledge_api_v1_knowledge_upload_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KnowledgeAcceptanceReportData": {
        "properties": {
          "answered_questions": {
            "default": 0,
            "title": "Answered Questions",
            "type": "integer"
          },
          "average_top_score": {
            "default": 0.0,
            "title": "Average Top Score",
            "type": "number"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeAcceptanceReportItemData"
            },
            "title": "Items",
            "type": "array"
          },
          "markdown": {
            "default": "",
            "title": "Markdown",
            "type": "string"
          },
          "missing_questions": {
            "default": 0,
            "title": "Missing Questions",
            "type": "integer"
          },
          "needs_review": {
            "default": false,
            "title": "Needs Review",
            "type": "boolean"
          },
          "total_questions": {
            "default": 0,
            "title": "Total Questions",
            "type": "integer"
          },
          "trusted_result_count": {
            "default": 0,
            "title": "Trusted Result Count",
            "type": "integer"
          }
        },
        "title": "KnowledgeAcceptanceReportData",
        "type": "object"
      },
...
      "KnowledgeAcceptanceReportRequest": {
        "additionalProperties": false,
        "properties": {
          "limit": {
            "default": 3,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "min_top_score": {
            "default": 0.7,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Top Score",
            "type": "number"
          },
          "questions": {
            "items": {
              "type": "string"
            },
            "maxItems": 50,
            "minItems": 1,
            "title": "Questions",
            "type": "array"
          }
        },
        "required": [
          "questions"
        ],
        "title": "KnowledgeAcceptanceReportRequest",
        "type": "object"
      },
...
      "KnowledgeImportRequest": {
        "additionalProperties": false,
        "properties": {
          "file_paths": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "File Paths",
            "type": "array"
          }
        },
        "required": [
          "file_paths"
        ],
        "title": "KnowledgeImportRequest",
        "type": "object"
      },
...
      "KnowledgeNormalizeConfirmRequest": {
        "additionalProperties": false,
        "properties": {
          "preview": {
            "$ref": "#/components/schemas/KnowledgeNormalizePreviewPayloadRequest"
          },
          "source": {
            "default": "",
            "maxLength": 500,
            "title": "Source",
            "type": "string"
          },
          "title": {
            "default": "",
            "maxLength": 200,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "preview"
        ],
        "title": "KnowledgeNormalizeConfirmRequest",
        "type": "object"
      },
      "KnowledgeNormalizeConfirmResultData": {
        "properties": {
          "file_name": {
            "default": "",
            "title": "File Name",
            "type": "string"
          },
          "file_path": {
            "default": "",
            "title": "File Path",
            "type": "string"
          },
          "import_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KnowledgeImportResultData"
              },
              {
                "additionalProperties": true,
                "type": "object"
              }
            ],
            "title": "Import Result"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          }
        },
        "title": "KnowledgeNormalizeConfirmResultData",
        "type": "object"
      },
...
      "KnowledgeNormalizePreviewPayloadRequest": {
        "additionalProperties": false,
        "properties": {
          "allowed_claims": {
            "items": {
              "type": "string"
            },
            "maxItems": 200,
            "title": "Allowed Claims",
            "type": "array"
          },
          "faq_items": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeNormalizeFaqItemRequest"
            },
            "maxItems": 100,
            "title": "Faq Items",
            "type": "array"
          },
          "forbidden_claims": {
            "items": {
              "type": "string"
            },
            "maxItems": 200,
            "title": "Forbidden Claims",
            "type": "array"
          },
          "handoff_rules": {
            "items": {
              "type": "string"
            },
            "maxItems": 200,
            "title": "Handoff Rules",
            "type": "array"
          },
          "source_excerpt": {
            "default": "",
            "maxLength": 8000,
            "title": "Source Excerpt",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "title": "Warnings",
            "type": "array"
          }
        },
        "title": "KnowledgeNormalizePreviewPayloadRequest",
        "type": "object"
      },
      "KnowledgeNormalizePreviewRequest": {
        "additionalProperties": false,
        "properties": {
          "source": {
            "default": "",
            "maxLength": 500,
            "title": "Source",
            "type": "string"
          },
          "text": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          },
          "title": {
            "default": "",
            "maxLength": 200,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "KnowledgeNormalizePreviewRequest",
        "type": "object"
      },
...
      "KnowledgeTrustedRebuildRequest": {
        "additionalProperties": false,
        "properties": {
          "acceptance_query": {
            "default": "",
            "maxLength": 500,
            "title": "Acceptance Query",
            "type": "string"
          }
        },
        "title": "KnowledgeTrustedRebuildRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WebKnowledgeBuildRequest": {
        "additionalProperties": false,
        "properties": {
          "file_paths": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "File Paths",
            "type": "array"
          },
          "search_limit": {
            "default": 5,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Search Limit",
            "type": "integer"
          }
        },
        "required": [
          "file_paths"
        ],
        "title": "WebKnowledgeBuildRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "WeChat AI Local API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/knowledge/acceptance-report": {
      "post": {
        "operationId": "acceptance_report_api_v1_knowledge_acceptance_report_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeAcceptanceReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_KnowledgeAcceptanceReportData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Acceptance Report",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/v1/knowledge/ai-normalize-confirm": {
      "post": {
        "operationId": "ai_normalize_confirm_api_v1_knowledge_ai_normalize_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeNormalizeConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_KnowledgeNormalizeConfirmResultData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ai Normalize Confirm",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/v1/knowledge/ai-normalize-preview": {
      "post": {
        "operationId": "ai_normalize_preview_api_v1_knowledge_ai_normalize_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeNormalizePreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_KnowledgeNormalizePreviewData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ai Normalize Preview",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/v1/knowledge/import": {
      "post": {
        "operationId": "import_knowledge_api_v1_knowledge_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_KnowledgeImportResultData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Knowledge",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/v1/knowledge/search": {
      "get": {
        "operationId": "search_knowledge_api_v1_knowledge_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 3,
              "maximum": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_KnowledgeSearchResultData__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Knowledge",
        "tags": [
          "knowledge"
        ]
      }
    },
...
    "/api/v1/knowledge/tasks": {
      "get": {
        "operationId": "list_knowledge_tasks_api_v1_knowledge_tasks_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_KnowledgeTaskData__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Knowledge Tasks",
        "tags": [
          "knowledge"
        ]
      }
    },
...
    "/api/v1/knowledge/trusted-rebuild": {
      "post": {
        "operationId": "rebuild_knowledge_with_trusted_embeddings_api_v1_knowledge_trusted_rebuild_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeTrustedRebuildRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_KnowledgeTrustedRebuildResultData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rebuild Knowledge With Trusted Embeddings",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/v1/knowledge/upload": {
      "post": {
        "operationId": "upload_knowledge_api_v1_knowledge_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_knowledge_api_v1_knowledge_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_KnowledgeImportResultData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Knowledge",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/v1/knowledge/web-build": {
      "post": {
        "operationId": "build_web_knowledge_api_v1_knowledge_web_build_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebKnowledgeBuildRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_WebKnowledgeBuildResultData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Build Web Knowledge",
        "tags": [
          "knowledge"
        ]
      }
    },
...
  }
}